requirements/requirements_prod.txt
python-telegram-bot~=22.5
python-dotenv~=1.2.1